move StoreVmDatabase to blockchain
