+++ crates/l2/prover/tests/perf_zkvm.rs
@@ -2,9 +2,9 @@
 #![allow(clippy::unwrap_used)]
 use ethrex_blockchain::Blockchain;
 use ethrex_common::types::{Block, ELASTICITY_MULTIPLIER};
+use ethrex_l2::utils::prover::db::to_prover_db;
 use ethrex_prover_lib::execute;
 use ethrex_storage::{EngineType, Store};
-use ethrex_vm::Evm;
 use std::path::Path;
 use tracing::info;
 use zkvm_interface::io::ProgramInput;
@@ -64,7 +64,7 @@ async fn setup() -> (ProgramInput, Block) {
         .unwrap()
         .unwrap();
 
-    let db = Evm::to_prover_db(&store.clone(), &vec![block_to_prove.clone()])
+    let db = to_prover_db(&store.clone(), &vec![block_to_prove.clone()])
         .await
         .unwrap();
