+++ crates/l2/utils/prover/mod.rs
@@ -1,3 +1,4 @@
+pub mod db;
 pub mod errors;
 pub mod proving_systems;
 pub mod save_state;
